Modified the Execution address of Uefi BL
The execution address of the Uefi Bootloader is modified. The Linux
kernel execution address is at the start of the DRAM base address, so
the boot loader execution address is moved the to the higher address so
that if required, it can be reclaimed by the OS for dynamic allocation.

Replaced the RTC Library class in dsc file.

The RTC Library was dependent on the ARM RTC PL031
Made it independent of Arm RTC by replacing it with the
RTC Template Library.

Signed-off-by: Girish K S <ks.giri@samsung.com>